Functionality to separate out trends and cycles
Given a time scale nT for which the cycle length is an integer 'Cycle_length',
separate_trend_scaled computes and removes the quadratic trend, then scales
'ft' smoothly to 'nT' points, computes the average cycle, and removes it from
the data, and scales back to the original time scale.

recombine_trend_scaled reintroduces the trend and average cycle.

The reason that scaling is needed is, for example, that the 3-day-averaged ocean data
gives a non-integer annual cycle, even with the exactly 365-day year, introducing a
shift over time.